Add files via upload
This is the same code, which works in py2 perfectly and as intended, unlike the py3 variant as of now. Running this for even 5k gens on a simple image will reproduce a decent pointillist remake of the given image. To run, make sure a py2 env is set up and running, preferably 2.7,16